flask-app/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-jwt-extended==4.7.1
python-dotenv==0.19.0